backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
dj-rest-auth>=5.0
django-allauth>=0.54
dj-database-url>=2.0
psycopg2-binary>=2.9
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
openai>=1.0.0
